requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
celery-types==0.26.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-filter==25.2
django-stubs==6.0.1
django-stubs-ext==6.0.1
djangorestframework==3.17.0
djangorestframework-stubs==3.16.8
kombu==5.6.2
Markdown==3.10.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==7.4.0
Serializer==0.2.1
setuptools==82.0.1
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
